raft: use correct types for configuration

+++ src/v/redpanda/config/configuration.h
@@ -19,7 +19,7 @@ struct configuration final : public config_store {
     // Network
     property<socket_address> rpc_server;
     // Raft
-    property<int64_t> node_id;
+    property<int32_t> node_id;
     property<int32_t> seed_server_meta_topic_partitions;
     property<std::vector<seed_server>> seed_servers;
     property<int16_t> min_version;
